Refactor integration test helpers
