+++ test/integration/common/assertLspCommand.ts
@@ -3,7 +3,12 @@ import * as vscode from "vscode";
 import * as assert from "assert";
 
 import { Client, Action } from "../common/types";
-import { isArray, isDefined, rangeEqual, uriEqual } from "../helpers/misc";
+import {
+  isArray,
+  isDefined,
+  rangeEqual,
+  uriEqual,
+} from "../helpers/assertions";
 
 export async function assertLspCommand(client: Client, action: Action) {
   await ensureValidationOfDoc(client, action);
