+++ test/integration/helpers/sleep.ts
@@ -1,0 +1,5 @@
+"use strict";
+
+export async function sleep(ms: number): Promise<void> {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
