+++ test/integration/tests/definition/definition.test.ts
@@ -2,7 +2,7 @@ import * as path from "path";
 import { getClient } from "../../client";
 import { Client } from "../../common/types";
 import { assertLspCommand } from "../../common/assertLspCommand";
-import { getDocUri } from "../../helpers/misc";
+import { getDocUri } from "../../helpers/docPaths";
 
 suite("Single-file Navigation", function () {
   this.timeout(10000);
